Update the test server to use the test interface.
svn path=/trunk/kdesupport/strigi/; revision=790101

+++ src/daemon/dbus/testinterface.cpp
@@ -32,7 +32,7 @@ TestInterface::concat(const std::string& a, const std::string& b) {
 }
 bool
 TestInterface::upload(const std::vector<char>& upload) {
-    printf("got file of size %i uploaded\n", upload.size());
+    printf("got file of size %i uploaded\n", (int)upload.size());
     return true;
 }
 map<string, string>
